+++ src/SWP/Bundle/CoreBundle/Tests/Functional/MultipleWebsitesPublish.php
@@ -209,10 +209,14 @@ final class MultipleWebsitesPublish extends WebTestCase
                         [
                             'tenant' => '123abc',
                             'route' => 3,
+                            'fbia' => false,
+                            'published' => true,
                         ],
                         [
                             'tenant' => '678iop',
                             'route' => 4,
+                            'fbia' => false,
+                            'published' => true,
                         ],
                     ],
                 ],
@@ -309,10 +313,14 @@ final class MultipleWebsitesPublish extends WebTestCase
                         [
                             'tenant' => '123abc',
                             'route' => 3,
+                            'fbia' => false,
+                            'published' => true,
                         ],
                         [
                             'tenant' => '678iop',
                             'route' => 4,
+                            'fbia' => false,
+                            'published' => true,
                         ],
                     ],
                 ],