+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This changelog references the relevant changes (bug and security fixes) done in 
 To get the diff for a specific change, go to https://github.com/superdesk/web-publisher/commit/XXX where XXX is the change hash
 
 * 1.4.0
+ * improvement [#524] Allow to set published to false when publishing an article
  * feature [#503] Added push integration/notification about newly created package based on WebSocket
  * bug [#499] Fixed package preview when template is not set or does not exist
  * bug [#494] Fixed article's slug line so it can handle chinese chars